refactor(agent): tools organized into catalog and collection subpackages
The taxonomy's two spaces become the folder structure: tools/catalog/
holds everything over the definitions the platform ships (sources,
assets, connections), tools/collection/ everything over the org's
persisted instances (sources, destinations, connections — listing,
setup, health checks). The former flat sources module splits along
that line; lineage, scheduling, and analytics stay at the root as
operational domains. Tool function names are unchanged, so the
LLM-facing surface is identical.

By Digitl
