+++ packages/interloper-agent/src/interloper_agent/tools/catalog/__init__.py
@@ -1,0 +1,5 @@
+"""Tools over the catalog — the component definitions the platform ships."""
+
+from interloper_agent.tools.catalog import assets, connections, sources
+
+__all__ = ["assets", "connections", "sources"]
