+++ packages/interloper-agent/src/interloper_agent/tools/catalog/assets.py
