+++ packages/interloper-agent/src/interloper_agent/tools/catalog/connections.py
@@ -1,0 +1,42 @@
+"""Connection tools over the catalog — the connection definitions the platform ships."""
+
+from __future__ import annotations
+
+from typing import Any
+
+from google.adk.tools.tool_context import ToolContext
+from interloper.oauth import is_provider_configured
+
+from interloper_agent.context import get_catalog
+
+
+def list_catalog_connections(tool_context: ToolContext) -> dict[str, Any]:
+    """List the connection definitions available in the catalog.
+
+    Use this to pick the right definition before requesting setup. Each
+    entry notes its required config fields and whether OAuth sign-in is
+    supported by the definition (``oauth``) and usable in this deployment
+    (``oauth_available``) — when OAuth is not available, the user will have
+    to enter credentials manually in the setup form. For the connections the
+    organisation actually has, use ``list_connections`` instead.
+    """
+    try:
+        catalog = get_catalog()
+        results = []
+        for key, defn in catalog.items():
+            if defn.get("kind") != "connection":
+                continue
+            schema = defn.get("config_schema") or {}
+            oauth = schema.get("x-oauth")
+            results.append({
+                "key": key,
+                "name": defn.get("name", key),
+                "description": defn.get("description"),
+                "provider": defn.get("provider"),
+                "required_fields": schema.get("required", []),
+                "oauth": oauth is not None,
+                "oauth_available": is_provider_configured(oauth["provider"]) if oauth else False,
+            })
+        return {"status": "success", "count": len(results), "connections": results}
+    except Exception as e:
+        return {"status": "error", "error": str(e)}
