+++ packages/interloper-agent/src/interloper_agent/tools/collection/__init__.py
@@ -1,0 +1,5 @@
+"""Tools over the org's collection — the component instances actually set up."""
+
+from interloper_agent.tools.collection import connections, destinations, sources
+
+__all__ = ["connections", "destinations", "sources"]
